+++ CHANGES.md
@@ -1,4 +1,5 @@
-- Slightly reorganized code
+- Added templating logic
+- Reorganized code
 
 Init
 - Fresh start of EJOcore